+++ packages/element/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @glissade/element
 
+## 0.76.0-pre.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @glissade/scene@0.76.0-pre.0
+  - @glissade/backend-canvas2d@0.76.0-pre.0
+  - @glissade/player@0.76.0-pre.0
+  - @glissade/core@0.76.0-pre.0
+
 ## 0.75.0
 
 ### Patch Changes